feat: adicionado endpoint para buscar os cabeçalhos na venda paginada

+++ template/backend/src/Ambev.DeveloperEvaluation.Domain/Repositories/ISaleRepository.cs
@@ -12,5 +12,5 @@ public interface ISaleRepository
 
     Task<Sale?> GetByIdAsync(Guid id, CancellationToken cancellationToken = default);
 
-    Task<(IEnumerable<Sale> Sales, long TotalRows)> GetByPaginedFilterAsync(short page,short pageSize);
+    Task<(IEnumerable<Sale> Sales, long TotalRows)> GetByPaginedFilterAsync(int page,int pageSize);
 }
